+++ winpty/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 authors = ["Zac Pullar-Strecker <zacmps@gmail.com>"]
 license = "MIT"
 description = "Safe rust bindings for winpty"
+edition = "2018"
 
 [target.'cfg(windows)'.dependencies]
 winpty-sys = "0.4.3"
@@ -13,3 +14,9 @@ widestring = "0.4.0"
 [target.'cfg(windows)'.dev-dependencies]
 named_pipe = "0.3"
 winapi = { version = "0.3", features = ["winnt", "processthreadsapi"] }
+
+[target.'cfg(windows)'.build-dependencies]
+embed-resource = "1.1.4"
+tempfile = "3.0.4"
+reqwest = "0.9"
+zip = "0.5"
